+++ packages/entities/entities-consumers/src/components/AddConsumerModal.cy.ts
@@ -9,14 +9,13 @@ describe('<AddConsumerModal/>', () => {
     const configGroupKonnect: KonnectConsumerListConfig = {
       app: 'konnect',
       controlPlaneId: '1234-asdf-asdf-asdf',
-      apiBaseUrl: '/us/kong-api/konnect-api',
+      apiBaseUrl: '/us/kong-api',
       createRoute: 'create-consumer',
       getViewRoute: () => 'view-consumer',
       getEditRoute: () => 'edit-consumer',
       consumerGroupId: '5921d16a-3e1d-4936-b21f-e5cae587415c',
       consumerGroupName: 'Test Group',
     }
-
     const interceptFilters = (params?: {
       mockData?: Consumer[];
       alias?: string;
@@ -24,7 +23,7 @@ describe('<AddConsumerModal/>', () => {
       cy.intercept(
         {
           method: 'GET',
-          url: `${configGroupKonnect.apiBaseUrl}/api/runtime_groups/${configGroupKonnect.controlPlaneId}/consumers*`,
+          url: `${configGroupKonnect.apiBaseUrl}/v2/control-planes/${configGroupKonnect.controlPlaneId}/core-entities/consumers*`,
         },
         {
           statusCode: 200,
@@ -40,7 +39,7 @@ describe('<AddConsumerModal/>', () => {
       cy.intercept(
         {
           method: 'POST',
-          url: `${configGroupKonnect.apiBaseUrl}/api/runtime_groups/${configGroupKonnect.controlPlaneId}/consumer_groups/${configGroupKonnect.consumerGroupId}/consumers`,
+          url: `${configGroupKonnect.apiBaseUrl}/v2/control-planes/${configGroupKonnect.controlPlaneId}/core-entities/consumer_groups/${configGroupKonnect.consumerGroupId}/consumers`,
         },
         {
           statusCode: status,