+++ packages/entities/entities-gateway-services/src/components/GatewayServiceList.cy.ts
@@ -31,7 +31,7 @@ describe('<GatewayServiceList />', () => {
   const baseConfigKonnect: KonnectGatewayServiceListConfig = {
     app: 'konnect',
     controlPlaneId: '1234-abcd-ilove-cats',
-    apiBaseUrl: '/us/kong-api/konnect-api',
+    apiBaseUrl: '/us/kong-api',
     createRoute,
     getViewRoute: () => viewRoute,
     getEditRoute: () => editRoute,
@@ -59,7 +59,7 @@ describe('<GatewayServiceList />', () => {
       cy.intercept(
         {
           method: 'GET',
-          url: `${baseConfigKonnect.apiBaseUrl}/api/runtime_groups/${baseConfigKonnect.controlPlaneId}/services*`,
+          url: `${baseConfigKonnect.apiBaseUrl}/v2/control-planes/${baseConfigKonnect.controlPlaneId}/core-entities/services*`,
         },
         {
           statusCode: 200,
@@ -496,7 +496,7 @@ describe('<GatewayServiceList />', () => {
       cy.intercept(
         {
           method: 'GET',
-          url: `${baseConfigKonnect.apiBaseUrl}/api/runtime_groups/${baseConfigKonnect.controlPlaneId}/services*`,
+          url: `${baseConfigKonnect.apiBaseUrl}/v2/control-planes/${baseConfigKonnect.controlPlaneId}/core-entities/services*`,
         },
         {
           statusCode: 200,
@@ -515,7 +515,7 @@ describe('<GatewayServiceList />', () => {
       cy.intercept(
         {
           method: 'GET',
-          url: `${baseConfigKonnect.apiBaseUrl}/api/runtime_groups/${baseConfigKonnect.controlPlaneId}/services*`,
+          url: `${baseConfigKonnect.apiBaseUrl}/v2/control-planes/${baseConfigKonnect.controlPlaneId}/core-entities/services*`,
         },
         (req) => {
           const size = req.query.size ? Number(req.query.size) : 30
@@ -574,7 +574,7 @@ describe('<GatewayServiceList />', () => {
         cy.intercept(
           {
             method: 'GET',
-            url: `${baseConfigKonnect.apiBaseUrl}/api/runtime_groups/${baseConfigKonnect.controlPlaneId}/services*`,
+            url: `${baseConfigKonnect.apiBaseUrl}/v2/control-planes/${baseConfigKonnect.controlPlaneId}/core-entities/services*`,
           },
           {
             statusCode: 500,