+++ packages/entities/entities-plugins/src/components/PluginList.cy.ts
@@ -22,7 +22,7 @@ const configureDynamicOrderingRoute = 'configure-dynamic-ordering-route'
 const baseConfigKonnect: KonnectPluginListConfig = {
   app: 'konnect',
   controlPlaneId: '1234-abcd-cat-loves-everyone',
-  apiBaseUrl: '/us/kong-api/konnect-api',
+  apiBaseUrl: '/us/kong-api',
   createRoute,
   getViewRoute: () => viewRoute,
   getEditRoute: () => editRoute,
@@ -67,7 +67,7 @@ describe('<PluginList />', () => {
       cy.intercept(
         {
           method: 'GET',
-          url: `${baseConfigKonnect.apiBaseUrl}/api/runtime_groups/${baseConfigKonnect.controlPlaneId}/plugins*`,
+          url: `${baseConfigKonnect.apiBaseUrl}/v2/control-planes/${baseConfigKonnect.controlPlaneId}/core-entities/plugins*`,
         },
         {
           statusCode: 200,
@@ -729,7 +729,7 @@ describe('<PluginList />', () => {
       cy.intercept(
         {
           method: 'GET',
-          url: `${baseConfigKonnect.apiBaseUrl}/api/runtime_groups/${baseConfigKonnect.controlPlaneId}/plugins*`,
+          url: `${baseConfigKonnect.apiBaseUrl}/v2/control-planes/${baseConfigKonnect.controlPlaneId}/core-entities/plugins*`,
         },
         {
           statusCode: 200,
@@ -748,7 +748,7 @@ describe('<PluginList />', () => {
       cy.intercept(
         {
           method: 'GET',
-          url: `${baseConfigKonnect.apiBaseUrl}/api/runtime_groups/${baseConfigKonnect.controlPlaneId}/plugins*`,
+          url: `${baseConfigKonnect.apiBaseUrl}/v2/control-planes/${baseConfigKonnect.controlPlaneId}/core-entities/plugins*`,
         },
         (req) => {
           const size = req.query.size ? Number(req.query.size) : 30
@@ -807,7 +807,7 @@ describe('<PluginList />', () => {
         cy.intercept(
           {
             method: 'GET',
-            url: `${baseConfigKonnect.apiBaseUrl}/api/runtime_groups/${baseConfigKonnect.controlPlaneId}/plugins*`,
+            url: `${baseConfigKonnect.apiBaseUrl}/v2/control-planes/${baseConfigKonnect.controlPlaneId}/core-entities/plugins*`,
           },
           {
             statusCode: 500,