+++ packages/entities/entities-shared/src/components/entity-filter/EntityFilter.vue
@@ -292,7 +292,6 @@ const applyFields = (hideMenu = false) => {
   :deep(.k-input) {
     padding-bottom: 4px!important;
     padding-top: 4px!important;
-    width: auto;
   }
 
   :deep(.k-input-wrapper) {