+++ packages/analytics/metric-cards/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.6](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/metric-cards@0.6.5...@kong-ui-public/metric-cards@0.6.6) (2023-11-30)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @kong/icons to ^1.8.1 ([#962](https://github.com/Kong/public-ui-components/issues/962)) ([66680c6](https://github.com/Kong/public-ui-components/commit/66680c6981424e3529478f8aec203de3ac4d2053))
+
+
+
+
+
+## [0.6.5](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/metric-cards@0.6.4...@kong-ui-public/metric-cards@0.6.5) (2023-11-30)
+
+**Note:** Version bump only for package @kong-ui-public/metric-cards
+
+
+
+
+
+## [0.6.4](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/metric-cards@0.6.3...@kong-ui-public/metric-cards@0.6.4) (2023-11-29)
+
+**Note:** Version bump only for package @kong-ui-public/metric-cards
+
+
+
+
+
+## [0.6.3](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/metric-cards@0.6.2...@kong-ui-public/metric-cards@0.6.3) (2023-11-28)
+
+**Note:** Version bump only for package @kong-ui-public/metric-cards
+
+
+
+
+
 ## [0.6.2](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/metric-cards@0.6.1...@kong-ui-public/metric-cards@0.6.2) (2023-11-22)